game.tsを分割して整理
