+++ packages/web/src/component/Emulator.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useRef } from "react";
-import Game from "./game";
-import type { Result } from "./game";
+import type { Result } from "./game/game";
+import TutorialGame from "./tutorial/tutorialGames";
+import Game from "./game/game";
 
 export interface User {
   name: string;
@@ -26,6 +27,7 @@ type EmulatorProps = {
   executionId: number; // エミュレーターそのものを更新するためのId
   handleStatuses: (statuses: Status[]) => void;
   onGameCompleted?: (result: Result) => void;
+  gameModeId: number;
 };
 
 export default function Emulator(props: EmulatorProps) {
@@ -38,19 +40,35 @@ export default function Emulator(props: EmulatorProps) {
     executionId,
     handleStatuses,
     onGameCompleted,
+    gameModeId,
   } = props;
   const canvasRef = useRef<HTMLCanvasElement>(null);
-  const gameRef = useRef<Game>();
+  const gameRef = useRef<any>();
   useEffect(() => {
     if (!canvasRef.current) throw new Error();
-    const game = new Game(users, canvasRef.current, (newStatuses: Status[]) => {
-      handleStatuses(newStatuses);
-    });
+    let GameClass;
+    switch (gameModeId) {
+      case 0:
+        GameClass = Game;
+        break;
+      case 1:
+        GameClass = TutorialGame;
+        break;
+      default:
+        GameClass = Game;
+    }
+    const game = new GameClass(
+      users,
+      canvasRef.current,
+      (newStatuses: Status[]) => {
+        handleStatuses(newStatuses);
+      }
+    );
     gameRef.current = game;
     return () => {
       game.destroy();
     };
-  }, [users, executionId, handleStatuses]);
+  }, [users, executionId, handleStatuses, gameModeId]);
   useEffect(() => {
     if (!gameRef.current) throw new Error();
     gameRef.current.onCompleted = (result: Result) => {
